feat: Add a base class named BaseAppTypeViewSet
